Delete postal-form files, consolidate billing-form directory
